+++ src/main/java/insane96mcp/iguanatweaksreborn/module/farming/feature/HarderCrops.java
@@ -1,24 +1,25 @@
 package insane96mcp.iguanatweaksreborn.module.farming.feature;
 
+import insane96mcp.iguanatweaksreborn.IguanaTweaksReborn;
 import insane96mcp.iguanatweaksreborn.module.Modules;
+import insane96mcp.iguanatweaksreborn.utils.Utils;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.LoadFeature;
-import insane96mcp.insanelib.util.IdTagMatcher;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TieredItem;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.CropBlock;
+import net.minecraft.world.level.block.state.BlockState;
 import net.minecraftforge.common.ToolActions;
 import net.minecraftforge.event.entity.player.PlayerEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.event.config.ModConfigEvent;
 import net.minecraftforge.registries.ForgeRegistries;
-
-import java.util.List;
 
 @Label(name = "Harder Crops", description = """
 		Crops are no longer insta-minable. This applies only to blocks that are instances of net.minecraft.world.level.block.CropBlock.
@@ -26,12 +27,10 @@ import java.util.List;
 		Changing anything requires a minecraft restart.""")
 @LoadFeature(module = Modules.Ids.FARMING)
 public class HarderCrops extends Feature {
+	public static final ResourceLocation HARDER_CROPS_TAG = new ResourceLocation(IguanaTweaksReborn.RESOURCE_PREFIX + "harder_crops");
 	@Config(min = 0d, max = 128d)
 	@Label(name = "Hardness", description = "How hard to break are plants? For comparison, dirt has an hardness of 0.5")
 	public static Double hardness = 1.0d;
-	@Config
-	@Label(name = "Other affected blocks", description = "Block ids or tags that will have the hardness and hoe efficiency applied. Each entry has a block or tag. This still only applies to blocks that have 0 hardness.")
-	public static List<IdTagMatcher> moreBlocksList = List.of();
 	@Config
 	@Label(name = "Only fully grown", description = "If the hardness should be applied to mature crops only.")
 	public static Boolean onlyFullyGrown = true;
@@ -54,18 +53,14 @@ public class HarderCrops extends Feature {
 		hardnessApplied = true;
 
 		for (Block block : ForgeRegistries.BLOCKS.getValues()) {
-			boolean isInWhitelist = false;
-			for (IdTagMatcher blacklistEntry : moreBlocksList) {
-				if (blacklistEntry.matchesBlock(block, null)) {
-					isInWhitelist = true;
-					break;
-				}
-			}
-			if (!(block instanceof CropBlock) && !isInWhitelist)
+			boolean isInTag = Utils.isBlockInTag(block, HARDER_CROPS_TAG);
+			if (!(block instanceof CropBlock) && !isInTag)
 				continue;
 			if (onlyFullyGrown) {
-				//I have doubts that this always takes the fully grown crop
-				block.getStateDefinition().getPossibleStates().get(block.getStateDefinition().getPossibleStates().size() - 1).destroySpeed = hardness.floatValue();
+				//I have doubts that this always takes the fully grown modded crops
+				BlockState state = block.getStateDefinition().getPossibleStates().get(block.getStateDefinition().getPossibleStates().size() - 1);
+				if (state.destroySpeed == 0f)
+					state.destroySpeed = hardness.floatValue();
 			}
 			else {
 				block.getStateDefinition().getPossibleStates().forEach(blockState -> {
@@ -86,14 +81,8 @@ public class HarderCrops extends Feature {
 		if (!heldItem.canPerformAction(heldStack, ToolActions.HOE_DIG) && !heldItem.canPerformAction(heldStack, ToolActions.AXE_DIG))
 			return;
 		Block block = event.getState().getBlock();
-		boolean isInWhitelist = false;
-		for (IdTagMatcher blacklistEntry : moreBlocksList) {
-			if (blacklistEntry.matchesBlock(block, null)) {
-				isInWhitelist = true;
-				break;
-			}
-		}
-		if (!(block instanceof CropBlock) && !isInWhitelist)
+		boolean isInTag = Utils.isBlockInTag(block, HARDER_CROPS_TAG);
+		if (!(block instanceof CropBlock) && !isInTag)
 			return;
 		float efficiency = heldItem.getTier().getSpeed();
 		if (efficiency > 1.0F) {